refactor: inline header qt/guiutil_font.h to guiutil.h
