+++ src/qt/addresstablemodel.cpp
@@ -6,7 +6,6 @@
 #include <qt/addresstablemodel.h>
 
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 #include <qt/walletmodel.h>
 
 #include <key_io.h>
