+++ src/qt/appearancewidget.cpp
@@ -10,7 +10,6 @@
 
 #include <qt/appearancewidget.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 #include <qt/optionsmodel.h>
 
 #include <util/system.h>
