+++ src/qt/askpassphrasedialog.cpp
@@ -12,7 +12,6 @@
 
 #include <qt/guiconstants.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 #include <qt/walletmodel.h>
 
 #include <support/allocators/secure.h>
