+++ src/qt/modaloverlay.cpp
@@ -7,7 +7,6 @@
 
 #include <chainparams.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 
 #include <QEasingCurve>
 #include <QResizeEvent>
