+++ src/qt/qrdialog.cpp
@@ -8,7 +8,6 @@
 #include <qt/bitcoinunits.h>
 #include <qt/guiconstants.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 #include <qt/qrimagewidget.h>
 
 #if defined(HAVE_CONFIG_H)
