+++ src/qt/qrimagewidget.cpp
@@ -5,7 +5,6 @@
 #include <qt/qrimagewidget.h>
 
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 
 #include <QApplication>
 #include <QClipboard>
