+++ src/qt/receivecoinsdialog.cpp
@@ -7,7 +7,6 @@
 
 #include <qt/addresstablemodel.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 #include <qt/optionsmodel.h>
 #include <qt/receiverequestdialog.h>
 #include <qt/recentrequeststablemodel.h>
