+++ src/qt/signverifymessagedialog.cpp
@@ -8,7 +8,6 @@
 
 #include <qt/addressbookpage.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 #include <qt/walletmodel.h>
 
 #include <key_io.h>
