+++ src/qt/trafficgraphwidget.cpp
@@ -6,7 +6,6 @@
 #include <qt/trafficgraphwidget.h>
 #include <qt/clientmodel.h>
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 
 #include <QPainter>
 #include <QPainterPath>
