+++ src/qt/utilitydialog.cpp
@@ -12,7 +12,6 @@
 #include <qt/forms/ui_helpmessagedialog.h>
 
 #include <qt/guiutil.h>
-#include <qt/guiutil_font.h>
 
 #include <clientversion.h>
 #include <init.h>
