+++ test/lint/lint-circular-dependencies.py
@@ -51,7 +51,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES = (
     "qt/bitcoingui -> qt/guiutil -> qt/bitcoingui",
     "qt/clientfeeds -> qt/clientmodel -> qt/clientfeeds",
     "qt/guiutil -> qt/qvalidatedlineedit -> qt/guiutil",
-    "qt/guiutil -> qt/guiutil_font -> qt/guiutil",
     "wallet/coinjoin -> wallet/receive -> wallet/coinjoin",
 )
